requirements.txt
streamlit
pycaret[full]
pandas-profiling
streamlit-pandas-profiling
plotly